requirements.txt
torch
torchvision
ultralytics
numpy
matplotlib
tqdm
Pillow
pandas

########################################
#      Optional / Needed for parts     #
########################################

# unik3d  # for UNIK3D as depth estimation method
# transformers  # for depth anything as depth estimation method
# waymo-open-dataset-tf-2-12-0  # for waymo dataset
# open3d  # visulaization